% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_distributions.R
\name{nb_catch}
\alias{nb_catch}
\alias{pois_catch}
\title{Tools for defining catch-size distributions}
\usage{
nb_catch(mean, variance)

pois_catch(mean)
}
\arguments{
\item{mean}{numeric The mean number of units per cluster (catch)}

\item{variance}{numeric The variance of the number of units per cluster
(catch)}
}
\value{
An object of class `distr` summarising the distribution of
  cluster/catch sizes
}
\description{
`nb_catch()` and `pois_catch()` are convenience functions for defining the
distribution of catche sizes (aka cluster sizes) for the negative binomial
and Poisson distributions respectively
}
\examples{
nb_catch(10,20)

pois_catch(10)
}
