% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input2.R
\name{check_geq2}
\alias{check_geq2}
\alias{check_in_range2}
\title{Check input helpers}
\usage{
check_geq2(val, min, allow_na = FALSE)

check_in_range2(val, allow_na = FALSE)
}
\arguments{
\item{val}{numeric/NA Input parameter}

\item{min}{numeric for check_geq2}

\item{allow_na}{boolean whether NAs are acceptable inputs or not. i.e. for 
clustering cases correlation/cost_cluster are NA}
}
\description{
Helper functions to validate input parameters.
}
