% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detection_errors}
\alias{detection_errors}
\alias{detection_errors.fixed_design}
\title{Power and sample size calculations for estimating population prevalence from
pooled samples}
\usage{
detection_errors(
  x,
  cluster_number,
  prevalence,
  correlation = 0,
  form = "logitnorm"
)

\method{detection_errors}{fixed_design}(
  x,
  cluster_number,
  prevalence,
  correlation = 0,
  form = "logitnorm"
)
}
\arguments{
\item{x}{a sample_design object}

\item{cluster_number}{numeric The number of clusters. The same sample design,
x, is assumed to be used at each cluster}

\item{prevalence}{numeric The proportion of units that carry the marker of
interest (i.e. true positive) that you want wish to assess your survey
against}

\item{correlation}{numeric The correlation between test results within a
single cluster (units in different clusters are assumed to be
uncorrelated). Must be a numeric value between 0 and 1, inclusive of both.
A value of 1 indicates that units within clusters are perfectly correlated
(there are no differences units within a single cluster). A value of 0
indicates that units within clusters are no more correlated than units in
different clusters.}

\item{form}{string The distribution used to model the cluster-level
prevalence and correlation of units within cluster. Select one of "beta",
"logitnorm" or "cloglognorm". See details.}

\item{sensitivity}{numeric The probability that the test correctly identifies
a true positive. Must be a numeric value between 0 and 1, inclusive of
both. A value of 1 indicates that the test can perfectly identify all true
positives.}

\item{specificity}{numeric The probability that the test correctly identifies
a true negative. Must be a numeric value between 0 and 1, inclusive of
both. A value of 1 indicates that the test can perfectly identify all true
negatives.}
}
\value{
A list with the typeI (false detection probability) and typeII error
  (false non-detection or 1-power of detection)
}
\description{
`detection_errors()` calculates the typeI (false detection probability) and
typeII error (false non-detection or 1-power of detection) for pool-tested
surveys with a known number and size of pools.
}
\examples{


detection_errors(fixed_design(10, 2), 20,
                 prevalence = 0.01,
                 sensitivity = 1, specificity = 0.999)

}
