% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_information.R
\name{fi_pool}
\alias{fi_pool}
\alias{fi_pool_cluster}
\alias{fi_pool_cluster_random}
\alias{fi_pool_random}
\title{Fisher information for population prevalence (and intra-cluster correlation)}
\usage{
fi_pool(pool_size, prevalence, sensitivity, specificity)

fi_pool_cluster(
  pool_size,
  pool_number,
  prevalence,
  correlation,
  sensitivity,
  specificity,
  form = "beta",
  real_scale = FALSE
)

fi_pool_cluster_random(
  catch_dist,
  pool_strat,
  prevalence,
  correlation,
  sensitivity,
  specificity,
  form = "beta",
  real_scale = FALSE,
  max_iter = 200,
  rel_tol = 1e-04
)

fi_pool_random(
  catch_dist,
  pool_strat,
  prevalence,
  sensitivity,
  specificity,
  max_iter = 200,
  rel_tol = 1e-04
)
}
\arguments{
\item{pool_size}{numeric The number of units per pool. Must be a numeric
value greater than or equal to 0.}

\item{prevalence}{numeric The proportion of units that carry the marker of
interest (i.e. true positive). Must be be a numeric value between 0 and 1,
inclusive of both.}

\item{sensitivity}{numeric The probability that the test correctly identifies
a true positive. Must be a numeric value between 0 and 1, inclusive of
both. A value of 1 indicates that the test can perfectly identify all true
positives.}

\item{specificity}{numeric The probability that the test correctly identifies
a true negative. Must be a numeric value between 0 and 1, inclusive of
both. A value of 1 indicates that the test can perfectly identify all true
negatives.}

\item{pool_number}{numeric The number of pools per cluster. Must be a numeric
value greater than or equal to 0.}

\item{correlation}{numeric The correlation between test results within a
single cluster (units in different clusters are assumed to be
uncorrelated). Must be a numeric value between 0 and 1, inclusive of both.
A value of 1 indicates that units within clusters are perfectly correlated
(there are no differences units within a single cluster). A value of 0
indicates that units within clusters are no more correlated than units in
different clusters.}

\item{form}{string The distribution used to model the cluster-level
prevalence and correlation of units within cluster. Select one of "beta",
"logitnorm" or "cloglognorm". See details.}

\item{real_scale}{boolean Applies only when `form = "logitnorm"` or `form =
"cloglognorm"`. Determines whether Fisher information should be returned
for the parameters of the logitnorm/cloglognorm distributions on the real
scale (i.e. mu and sigma). If FALSE (the default) Fisher information is
returned for prevalence (theta) and correlation (rho) instead.}
}
\value{
The Fisher information for prevalence (`fi_pool()`) or the Fisher
  information matrix for prevalence and intra-cluster correlation
  (`fi_pool_cluster()` and `fi_pool_cluster_random()`)
}
\description{
`fi_pool()` and `fi_pool_cluster()` calculate Fisher information (FI) for
pool-tested surveys with a known number and size of pools. `fi_pool_random()`
and `fi_pool_cluster_random()` calculate FI for surveys where the number of
units is random. `fi_pool()` and `fi_pool_random` calculates the Fisher
information for the prevalence for simple random surveys.
`fi_pool_cluster()` and `fi_pool_cluster_random()` calculate the two-by-two
Fisher information matrix for prevalence and within-cluster correlation for
cluster survey designs.
}
\examples{
fi_pool(
  pool_size = 10, prevalence = 0.01, sensitivity = 0.95, specificity = 0.99
  )

fi_pool_cluster(
  pool_size = 10, pool_number = 5, prevalence = 0.01,
  correlation = 0.05, sensitivity = 0.95, specificity = 0.99
  )


}
