% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_design.R
\name{fixed_design}
\alias{fixed_design}
\title{fixed_design s3 constructor}
\usage{
fixed_design(
  pool_size = NULL,
  pool_number = NULL,
  sensitivity = 1,
  specificity = 1
)
}
\arguments{
\item{pool_size}{numeric/NULL The number of units per pool. Must be a numeric
value greater than 0.}

\item{pool_number}{numeric/NULL The number of pools per cluster. Numeric
inputs must be an integer greater than or equal to 1.}

\item{sensitivity}{numeric The probability that the test correctly identifies
a true positive. Must be a numeric value between 0 and 1, inclusive of
both. A value of 1 indicates that the test can perfectly identify all true
positives.}

\item{specificity}{numeric The probability that the test correctly identifies
a true negative. Must be a numeric value between 0 and 1, inclusive of
both. A value of 1 indicates that the test can perfectly identify all true
negatives.}
}
\value{
An object of class \code{sample_design}
}
\description{
Stores parameters related to the sampling design. Aims to reduce having to
input each param separately across functions (e.g. power/optimise).
}
\examples{
perfect <- fixed_design(pool_size = 10)

imperfect <- fixed_design(
  pool_size = 10, pool_number = NULL, sensitivity = 0.95, specificity = 0.99
)
}
