% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_prevalence.R
\name{optimise_prevalence}
\alias{optimise_prevalence}
\alias{optimise_prevalence.fixed_design_optimise_sN}
\title{Optimise prevalence}
\usage{
optimise_prevalence(x, ...)

\method{optimise_prevalence}{fixed_design_optimise_sN}(
  x,
  prevalence,
  cost_unit,
  cost_pool,
  cost_cluster,
  correlation,
  max_s = 50,
  max_N = 20,
  form = "logitnorm",
  ...
)
}
\arguments{
\item{x}{a sample_design object}

\item{...}{Additional arguments passed to methods.}

\item{prevalence}{numeric The proportion of units that carry the marker of
interest (i.e. true positive). Must be be a numeric value between 0 and 1,
inclusive of both.}

\item{cost_unit}{numeric The cost to process a single unit. Must be a numeric
value greater than or equal to 0.}

\item{cost_pool}{numeric The cost to process a single pool. Must be a numeric
value greater than or equal to 0.}

\item{cost_cluster}{numeric/NA The cost to process a cluster. Must be a
numeric value greater than or equal to 0. For `optimise_s_prevalence()`,
this can be ignored if correlation is NA.}

\item{correlation}{numeric The correlation between test results within a
single cluster (units in different clusters are assumed to be
uncorrelated). Must be a numeric value between 0 and 1, inclusive of both.
A value of 1 indicates that units within clusters are perfectly correlated
(there are no differences units within a single cluster). A value of 0
indicates that units within clusters are no more correlated than units in
different clusters.}

\item{max_s}{numeric The maximum number of units per pool (pool size).}

\item{max_N}{numeric The maximum number of pools per cluster (pool number).}

\item{form}{string The distribution used to model the cluster-level
prevalence and correlation of units within cluster. Select one of "beta",
"logitnorm" or "cloglognorm".}
}
\description{
Placeholder documentation to track params for the refactor.
}
