% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_strat_methods.R
\name{print.pool_strat}
\alias{print.pool_strat}
\title{Methods for pool_strat obejcts}
\usage{
\method{print}{pool_strat}(x, ...)
}
\arguments{
\item{x}{An object of class `pool_strat` (typically produced by one of the
built-in functions for generating them such as `pool_max_size()` and
`pool_taget_number()`}
}
\description{
Methods for pool_strat obejcts
}
