% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling_strategies.R
\name{pool_max_size}
\alias{pool_max_size}
\alias{pool_target_number}
\title{Tools for defining pooling strategies}
\usage{
pool_max_size(max_size)

pool_target_number(target_number)
}
\arguments{
\item{max_size}{integer The maximum number of units per pool}

\item{target_number}{integer The variance of the number of units per cluster
(catch)}
}
\value{
A function of class `pool_strat` with a single argument, `catch`, for
  the catch size that returns a list of the pool sizes and numbers that the
  catch should be split into
}
\description{
`pool_max_size()` and `pool_taget_number()` are functions for defining the how a
given number of units should be divided into pools. Using `pool_max_size()`
pools have fixed maximum pool size (`max_size`) with any remainder placed in
a single smaller pool. Using `pool_target_number()` pools have a target number (target)
of approximately equal sized pools, with fewer pools used only if
}
\examples{
pool_max_size(10)

pool_target_number(4)
}
