% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_utils.R
\name{power_size_results}
\alias{power_size_results}
\title{S3 Class and Methods for Power and Sample Size Calculation Outputs}
\usage{
power_size_results(
  pool_size,
  pool_number,
  cluster_number,
  prev_null,
  prev_alt,
  sig_level,
  power,
  alternative,
  target
)
}
\arguments{
\item{...}{Input and inferred variables from relevant functions.}
}
\value{
An object of class \code{power_size_results} containing selected
  input parameters and results.
}
\description{
As many variables are shared across groups of functions (e.g. `power_pool`
and `sample_size_pool`), this class and print method ensures that outputs are
stored and displayed consistently, summarising calculations results and the
inputs for context.
}
\examples{
# For power_pool()
result <- power_size_results(pool_size = pool_size, pool_number = pool_number, cluster_number = cluster_number,
                             prev_null = theta0, prev_alt = theta0, sig_level = sig_level,
                             power = power, alternative = alternative, target = "power")
print(result)
}
