% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_threshold}
\alias{power_threshold}
\alias{sample_size_threshold}
\title{Power and sample size calculations for detecting presence and estimating
population prevalence from pooled samples}
\usage{
power_threshold(
  x,
  cluster_number,
  prevalence_null,
  prevalence_alt,
  correlation,
  sig_level,
  alternative,
  form,
  link,
  ...
)

sample_size_threshold(
  x,
  prevalence_null,
  prevalence_alt,
  correlation,
  power,
  sig_level,
  alternative,
  form,
  link,
  ...
)
}
\arguments{
\item{x}{sample_design object.}

\item{cluster_number}{numeric The total number of clusters in a cluster
survey design (should be greater than 1) or 1 surveys where all collection
happens at a single site or collected via simple random sampling from the
target population.}

\item{prevalence_null, prevalence_alt, prevalence}{numeric The proportion of
units that carry the marker of interest (i.e. true positive). For
`_threshold` functions `prevalence_null` is the threshold to compare to and
`prevalence_alt` is the design prevalence. For `_detect` functions,
`prevalence` is the design prevalence. Must be be a numeric value between 0
and 1, inclusive of both.}

\item{correlation}{numeric The correlation between test results within a
single cluster (units in different clusters are assumed to be
uncorrelated). Must be a numeric value between 0 and 1, inclusive of both.
A value of 1 indicates that units within clusters are perfectly correlated
(there are no differences units within a single cluster). A value of 0
indicates that units within clusters are no more correlated than units in
different clusters.}

\item{sig_level}{numeric Significance level for statistical test. Defaults to
0.05. Must be strictly between 0 and 1.}

\item{alternative}{string The kind of comparison to make. If "greater" or
"less" (default), computes power of tests for one-sided comparisons that
population prevalence (`prevalence_alt`) is greater than or less than the
reference on threshold prevalence (`prevalence_null`) respectively. If
"two.sided" computes power for a two-sided comparison (`power_threshold()`
only).}

\item{form}{string The distribution used to model the cluster-level
prevalence and correlation of units within cluster. Select one of "beta",
"logitnorm" or "cloglognorm". See details.}

\item{link}{string Transformation to be applied to prevalence estimates for
the purposes of calculating confidence intervals. Options are 'identity'
(i.e. no transformation), 'logit' (default), 'cloglog' and 'log'.}

\item{...}{Additional parameters.}

\item{power}{numeric The desired statistical power of the survey.}

\item{max_iter}{numeric Maximum number of iterations (possible catch sizes)
to consider when calculating expected FI over random catch sizes. Generally
needs to be large enough so that the nearly all catch sizes will be less
than `max_iter` otherwise algorithm will terminate early (with a warning)}

\item{rel_tol}{numeric Relative tolerance for determining convergence when
calculating expected FI over random catch sizes. Must be positive and
should be much smaller than 1.}
}
\value{
The statistical power of the proposed design with regards to
  comparing prevalence to a threshold (`power_threshold()`) or a list with
  the sample size (number of clusters, pools, and units) required to achieve
  desired power (`sample_size_threshold()`)
}
\description{
`power_threshold()` calculates the statistical power of a pooled survey
design to determine whether population prevalence is different from a
threshold.
}
\details{
`power_detect()` calculations the statistical power of a pooled survey design
to detect a marker in a population, given it is present at a specified
prevalence.

`sample_size_` family of functions calculate the sample size required for a
pooled survey to achieve the specified power for the given objective.
}
\examples{
# Fixed design examples ----
fd <- fixed_design(pool_size = 10, pool_number = 2)

## Unclustered ----
power_threshold(fd, cluster_number = 50,
           prevalence_null = 0.02, prevalence_alt = 0.1)

sample_size_threshold(fd, prevalence_null = 0.02, prevalence_alt = 0.1)

## Clustered ----
power_threshold(fd, cluster_number = 50,
                prevalence_null = 0.02, prevalence_alt = 0.01,
                correlation = 0.1)

sample_size_threshold(fd,
                      prevalence_null = 0.02, prevalence_alt = 0.01,
                      correlation = 0.1)

# Variable design examples ----

vd <- variable_design(nb_catch(20,25), pool_target_number(2))
power_threshold(vd, cluster_number = 50,
                prevalence_null = 0.02, prevalence_alt = 0.01,
                correlation = 0.1)

sample_size_threshold(vd,
                      prevalence_null = 0.02, prevalence_alt = 0.01,
                      correlation = 0.1)
}
