% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_fisher_information.R
\name{sample_design}
\alias{sample_design}
\title{sample_design constructor}
\usage{
sample_design(pool_size, pool_number = NULL, sensitivity = 1, specificity = 1)
}
\arguments{
\item{pool_size}{}

\item{pool_number}{}

\item{sensitivity}{}

\item{specificity}{}
}
\value{
An object of class \code{sample_design}
}
\description{
sample_design constructor
}
\examples{
sd <- sample_design(pool_size = 10, pool_number = NULL, sensitivity = 1, specificity = 1)
fi_pool(sd)
}
