% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_information.R
\name{fi_pool}
\alias{fi_pool}
\alias{fi_pool_cluster}
\title{Fisher information for population prevalence}
\usage{
fi_pool(pool_size, prevalence, sensitivity, specificity)

fi_pool_cluster(
  pool_size,
  pool_number,
  prevalence,
  correlation,
  sensitivity,
  specificity,
  form = "beta",
  real_scale = FALSE
)
}
\arguments{
\item{pool_size}{numeric The number of units per pool. Must be a numeric
value greater than or equal to 0.}

\item{prevalence}{numeric The proportion of units that carry the marker of
interest (i.e. true positive). Must be be a numeric value between 0 and 1,
inclusive of both.}

\item{sensitivity}{numeric The probability that the test correctly identifies
a true positive. Must be a numeric value between 0 and 1, inclusive of
both. A value of 1 indicates that the test can perfectly identify all true
positives.}

\item{specificity}{numeric The probability that the test correctly identifies
a true negative. Must be a numeric value between 0 and 1, inclusive of
both. A value of 1 indicates that the test can perfectly identify all true
negatives.}

\item{pool_number}{numeric The number of pools per cluster. Must be a numeric
value greater than or equal to 0.}

\item{correlation}{numeric The correlation between test results within a
single cluster (units in different clusters are assumed to be
uncorrelated). Must be a numeric value between 0 and 1, inclusive of both.
A value of 1 indicates that units within clusters are perfectly correlated
(there are no differences units within a single cluster). A value of 0
indicates that units within clusters are no more correlated than units in
different clusters.}

\item{form}{string The distribution used to model the cluster-level
prevalence and correlation of units within cluster. Select one of "beta",
"logitnorm" or "cloglognorm". See details.}

\item{real_scale}{boolean Applies only when `form = "logitnorm"` or `form =
"cloglognorm"`. Determines whether Fisher information should be returned
for the parameters of the logitnorm/cloglognorm distributions on the real
scale (i.e. mu and sigma)? If FALSE (the default) Fisher information is
returned for prevalence (theta) and correlation (rho) instead.}
}
\value{
A numeric value of the Fisher information
}
\description{
`fi_pool()` and `fi_pool_cluster()` calculates the Fisher Information for
pool testing strategies for a given number and size of pools, where the
sensitivity and specificity of the test are known. `fi_pool()` calculates the
Fisher information for the prevalence for simple random surveys.
`fi_pool_cluster()` calculates the two-by-two Fisher information matrix for
prevalence and within-cluster correlation for cluster survey designs.
}
\examples{
fi_pool(
  pool_size = 10, prevalence = 0.95, sensitivity = 0.95, specificity = 0.99
  )

fi_pool_cluster(
  pool_size = 10, pool_number = 5, prevalence = 0.01,
  correlation = 0.05, sensitivity = 0.95, specificity = 0.99
  )
}
