% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_utils.R
\name{power_size_results}
\alias{power_size_results}
\title{S3 class for power and sample size calculation outputs with pooled tests}
\usage{
power_size_results(
  design,
  cluster_number,
  prev_null = NULL,
  prev_alt = NULL,
  prev = NULL,
  correlation,
  sig_level = NULL,
  power,
  alternative = NULL
)
}
\arguments{
\item{design}{required; sample design}

\item{cluster_number}{required}

\item{prev_null, prev_alt, prev}{either prev, or prev_null and prev_alt required}

\item{correlation}{required}

\item{sig_level}{}

\item{power}{required}

\item{alternative}{}
}
\value{
An object of class \code{power_size_results} containing selected
  input parameters and results.
}
\description{
As many variables are shared across groups of functions (e.g. `power_pool`
and `sample_size_pool`), this class and print method ensures that outputs are
stored and displayed consistently, summarising calculations results and the
inputs for context.
}
